services/document-processor/requirements.txt
doctr
ultralytics
pdf2image
fastapi
python-multipart
supabase-py
uvicorn
torch
opencv-python
numpy
pandas
python-dotenv